+++ backend/leavecalc/src/main/java/com/lawding/leavecalc/domain/global/config/DynamoDbConfig.java
@@ -1,6 +1,5 @@
 package com.lawding.leavecalc.domain.global.config;
 
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import software.amazon.awssdk.enhanced.dynamodb.DynamoDbEnhancedClient;
