--- backend/leavecalc/src/main/java/com/lawding/leavecalc/domain/stats/entity/DailyUser.java
@@ -1,26 +1,0 @@
-package com.lawding.leavecalc.domain.stats.entity;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
-import java.time.LocalDate;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Table(name = "daily_user")
-public class DailyUser {
-
-    @Id
-    @Column(name = "record_date")
-    private LocalDate recordDate;
-
-    private int web;
-    private int ios;
-    private int android;
-
-}
