+++ backend/leavecalc/src/main/java/com/lawding/leavecalc/domain/stats/repository/DailyUserRepository.java
@@ -1,10 +1,29 @@
 package com.lawding.leavecalc.domain.stats.repository;
 
-import com.lawding.leavecalc.domain.stats.entity.DailyUser;
+import com.lawding.leavecalc.domain.appversion.model.AppVersion;
+import com.lawding.leavecalc.domain.stats.model.DailyUser;
 import java.time.LocalDate;
-import org.springframework.data.jpa.repository.JpaRepository;
+import java.util.List;
+import java.util.Optional;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
+import software.amazon.awssdk.enhanced.dynamodb.DynamoDbEnhancedClient;
+import software.amazon.awssdk.enhanced.dynamodb.DynamoDbTable;
+import software.amazon.awssdk.enhanced.dynamodb.TableSchema;
 
 @Repository
-public interface DailyUserRepository extends JpaRepository<DailyUser, LocalDate> {
+public class DailyUserRepository {
+
+    private final DynamoDbTable<DailyUser> table;
+
+    private static final String TABLE_NAME = "app_version_policy";
+
+    @Autowired
+    public DailyUserRepository(DynamoDbEnhancedClient enhancedClient) {
+        this.table = enhancedClient.table(TABLE_NAME, TableSchema.fromBean(DailyUser.class));
+    }
+
+    public List<DailyUser> scanAll() {
+        return table.scan().items().stream().toList();
+    }
 }
