+++ backend/leavecalc/src/main/java/com/lawding/leavecalc/domain/stats/service/impl/UserStatsServiceImpl.java
@@ -17,7 +17,7 @@ public class UserStatsServiceImpl implements UserStatsService {
 
     @Override
     public List<DailyUserDto> getAllDailyUser() {
-        return dailyUserRepository.findAll().stream()
+        return dailyUserRepository.scanAll().stream()
             .map(DailyUserDto::from)
             .toList();
     }
